move industry components to a components folder
