+++ src/app/(public)/industry/components/IndustryComponent.tsx
